Add tracing experiment for SWI calls.
Being able to output the information about the SWI calls that are
being executed is very useful. The initial 'traceswi' debug option
provides just this information, together with the registers that
are present at the time of the SWI call.

The extra 'traceswiargs' uses information from a built in database
of the parameters that are present for certain SWIs. These parameters
have been extracted from OSLib, and provide details of the names of
the registers that are used by the SWIs that have been described. In
some cases, these SWIs cannot be described properly because there are
multiple variants of the SWIs present in OSLib (because that's how
the SWI works). In these cases, we have no entry in the database and
the result is the whole register set being reported.