Add exit parameters to swi tracing; improved oslib parser output.
The swi_conditions.py module is now output as structure with the
description and the exit registers in addition to the entry registers,
so that we can report the returns as well. The SWI return is now
reported to the trace module, so that it can report on the behaviour
as well. This means that we have a lot more verbose output, but it
should be significantly easier to see what's going on in the code.

The oslib parser code has better generation of the output, and can
now be given multiple files on the command line and will generate
the entire contents of swi_conditions.py from that.