Add support for variadic SWI decoding to the trace handling.
Variadic SWIs - those whose parameters change depending on the
reason supplied to them - are a harder area to try to decode when
printing the list of arguments to the calls. Conditional code has
been extracted from some of the definitions to allow us to produce
a list of potential arguments that can be used for each SWI. These
include a set of 'match' parameters that must be met for that set
of arguments to be used.

This appears to work pretty well on the simpler cases. There will
be times when the calls have a pattern that is difficult to handle,
and won't be handled. However, the 'traceswiargs' output has
improved now that we handle these variadic entries.