Add ability to create PyModule templates; fix merging SWI definitions.
Simple PyModules can now be created from the defmod definitions, allowing
a simpler bootstrap.

The SWI definitions that were in one file and continued in another would
not be populated correctly in the swi_conditions file output. We now merge
the duplicated SWIs so that they can be listed properly in the swi_conditions
file. The more files we process, the more useful that is.