Improve definition of the template pymodules from OSLib headers.
The python modules now generate the correct SWI base and name even
when def file contains an OS entry (eg vector, event service call).
These are now removed from the SWIs that are generated within the
module.

The comments now describe the conditions a more more clearly,
including marking the corrupted registers as such, and constants
with a more descriptive statement.

After the register read, we now write out a commented block of code
that would update the output registers as required. This should
make it easier to provide implementations.