Add a tentative API generator for the OSLib definitions.
The OSLib definitions can be converted to definitions that can be used
in the internal APIs within riscos.api. This should make it a lot easier
to create APIs for things that the pymodules wish to call.

It could be used to help Chris with his SWI interfaces for his port of
Python, too. Depends how easy it would be to implement the calls in
the style used by his Python call. Of course, it could in theory create
dedicated C code to handle those SWIs.