Add line number reports to the oslib parser code.
When something goes wrong with the parse, it only reported that it
was expecting to see one thing and got something else. There was no
indication of where that was. Line number reporting has been added
for the base line, so that it's easier to see where it's encountering
problems.