Update OSLib parser to understand the '*' on constant values.
The '*' on constant values means that the description of the SWI
applies to the constant value as a reason code. This was not understood
by the parser, so it would fail to read those files at all. The actual
support for understanding these description markers is not at all
supported - indeed the description and the star will just be thrown away
at present, but it's enough to make the constants and SWIs parse.