Some initial work to try to create some interfaces for RISC OS Python.
In order to try to create some structures for RISC OS python, we can
use the ctypes structures to access the data. It's pretty simplified
really because I'm not doing a lot of the things that the OSLib parser
did. But it looks like it's generating a number of useful structures
that look about right to my cursory glance.