Support for generating some interface methods for SWIs.
The interface methods are pretty simple, and not quite right yet.
Although we return the correct registers, we don't process the
returned tuple properly - we treat it as a list of registers.

Additionally, the swi function returns a value, not a tuple, when
there is only a single return value. This isn't handled at the
moment.