Fix the return parameters to return more usefully; build-all-apis.sh.
The return parameters should now be correct when we return. This has
involved a bit of messing around, but seems to be ok - even if the
Jinja2 is really hard to read. Some more comments have been added
around the place to make it easier to work with.

A simple script has been added which tries to obtain OSLib and then
process all the files through it. There's still a lot of errors reported
but it does quite well.