Add a tentative CI job that tries to build everything.
Might as well try to build the APIs as we might be able to use these
in the RISC OS Python (and in Pyromaniac).

Fixes for running under Python 3 - the CI runs in Python 3, so we
may as well fix the bits that weren't quite right. This is largely in
the encoding of the strings.