Add support for resolving 'needs' whilst reading defs.
We now resolve the 'needs' in each of the defs, so that we know about
all the constants and types that the def file repends on. We then
write out only the ones that we need for the files that we are
processing. This has the nice side effect that, given the oslib
directory, we can now parse the def files by just giving its name
instead of the actual path to it.