Add makefile to build all the different parts of the OSLib outputs.
The makefile should be able to generate all the output, which we can
use to check that the library is doing the right thing. In theory this
could be made into a CI test, but it's not working properly yet - the
API interface generation is currently erroring. As more code is copied
from the other sources we should get this working better.