Added rudimentary templates for creating C modules.
Creating C modules isn't that hard using the definitions we have.
This change introduces the CMHG file and the C source file which
can be used to start a module off quickly.

However, we can probably do a little better with the type of
parameters that are used within the interfaces. That can come later.