Updated the C module templates to create header with types and constants.
The C module templates are now able to create a header with the types
and constants that are needed for this particular module (all using the
OSLib naming) so that it can be worked on standalone. The modules can
almost all be built like this. There are some issues, but they're not
that significant - notable problems:

* ConvertSprite (reuse of the variable 'regs')
* DiagnosticDump (reuse of the variable 'regs')
* DrawFile (Draw_PathElement is referenced within Draw_PathElement)
* FileSwitch (wrong CMHG SWI chunk base)
* ImageFileConvert (misinterpreted Void type)
* ImageFileRender (misinterpreted .int type)
* InverseTable (misinterpreted OSSpriteOp_Area type - should be .Ref?)
* PDriver (misinterpreted .int type, Draw_PathElement referenced within Draw_PathElement)
* ZapFontMenu (wrong CMHG SWI chunk base)